% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{cdfDiff}
\alias{cdfDiff}
\title{Cumulative probability distribution segments}
\usage{
cdfDiff(a, b, dist, ...)
}
\arguments{
\item{a}{Vector of the left sides of segments. Must be the same length as
b=, or a single value. If a single value, that value will be used as the
left side of all segments}

\item{b}{Vector of the right sides of segments. Must be the same length as
a=, or a single value. If a single value, that value will be used as the
right side of all segments.}

\item{dist}{Name of distribution to use, one of: "normal". Will be partially
matched, case insensitively, but must match uniquely.}

\item{...}{Distribution parameter, e.g. mean= and sd= for the normal
distribution.}
}
\value{
The probability of each of the specified line segments. Note that
for segments specified backwards like \code{b[i] < a[i]}, the corresponding
probabilities will be negative.
}
\description{
Calculates the difference between two values of a cumulative probability
distribution. Will result in negative probabilities when cdf(b) < cdf(a),
e.g. when the ends of the cdf probability segment are reversed.
}
\examples{
# P( x > 0 ), standard normal distribution
cdfDiff( 0, Inf, "norm" )
#> [1] 0.5

# P( x > 0 ), standard normal distribution (backwards)
cdfDiff( Inf, 0, "norm" )
#> [1] -0.5

# P( x < 2.1 ), normal with mean= 2.1 ( and sd= 1)
cdfDiff( -Inf, 2.1, "norm", mean= 2.1 )
#> [1] 0.5

# P( x = -Inf, x < 0, x < Inf ), standard normal distribution
cdfDiff( -Inf, c(-Inf, 0, Inf), "norm" )
#> [1] 0.0 0.5 1.0

# Mix of forwards and backwards
cdfDiff( c(-Inf, -2.1, Inf), c(-2.1, -Inf, -2.1), "norm", mean=-2.1 )
#> [1]  0.5 -0.5 -0.5

# P( x < 2 or x > 2 ) standard normal
sum(cdfDiff( c(-Inf, 2), c(-2, Inf), "norm"))
#> 0.04550026
}
