% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{checkFileSum}
\alias{checkFileSum}
\title{Validate a file against a checksum file}
\usage{
checkFileSum(
  file,
  algo = c("md5", "sha1", "crc32", "sha256", "sha512"),
  checkFile = NULL,
  as = c("test", "check", "error", "warning", "message", "stop")
)
}
\arguments{
\item{file}{The file to validate}

\item{algo}{The checksum algorithm to use (see \code{digest::digest()}).}

\item{checkFile}{The name of the file holding the expected checksum value.
If \code{NULL}, will attempt to guess; assumes named the same as \verb{file=}, except
with an additional extension matching \verb{algo=}. If not found, looks for a
hidden checkFile (with a leading "."). e.g. an md5 checkFile for
\code{file.txt} would be guessed as \code{file.txt.md5} or \code{.file.txt.md5}.}

\item{as}{How the results of the checksum test are reported. \code{as="test"}
returns \code{TRUE} if the checksums match, \code{FALSE} if they do not, and \code{NA} if
there is an error looking for the requisite files. All others rely on a
generated string - either the empty string \code{""} when the checksums match,
an error with the file names and checksums when they fail to match, or an
error if something prevents them from being calculated (e.g. missing
files). \code{as="check"} returns this string, \code{as="error"} and \code{as="stop"} are
aliases; they either return the generated empty string (silently), or
trigger an error exit if the string is not empty. \code{as="warning"} and
\code{as="message"} return the generated string silently, but also signal the
appropriate message if the string is not empty.}
}
\value{
Depends on \verb{as=}. With matching checksums, will either be TRUE or the
empty string \verb{"} (possibly silently). If checksums don't match, it will
either be FALSE, a string giving the files and their mismatched checksums
(possibly silently), or an error will be triggered. If there is a problem
calculating the checksums, e.g. a file does not exits, the result will
either be "NA", a string describing the problem, or an error will be
triggered.
}
\description{
Validates a file using a specified checksum algorithm against a file that
contains just the expected checksum result. By default looks for a file
in parallel to the file tested, except with an additional extension matching
the checksum algorithm. Can specify what to do after testing match through
use of the \verb{as=} argument. Supports returning TRUE/FALSE, a string message,
or signaling a message via \code{stop()}, \code{warning()}, or \code{message()}. To support
testing without throwing errors, any errors such as missing file are reported
the same as an error triggered by a mismatched checksum, except it causes NA
to be returned with \code{as=test}.
}
