% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringUtils.R
\name{commonSuffix}
\alias{commonSuffix}
\title{Longest common suffix}
\usage{
commonSuffix(x, ignoreCase = FALSE, dropNA = FALSE)
}
\arguments{
\item{x}{The strings to find the longest common suffix for.}

\item{ignoreCase}{Set this true to match suffixes even if they differ in case.}

\item{dropNA}{Set this true to ignore \code{NA}s when searching for a common
suffix}
}
\value{
The common suffix, if any, or "", if no common suffix can be found.
}
\description{
Finds and returns the longest suffix common to all strings in a character
vector. Can be set to ignore case, in which case the returned common string
will be in lower case. If any strings are \code{NA}, this returns \code{NA}
unless \code{dropNA= TRUE} is set, which will drop \code{NA}s before
checking for a common suffix
}
\examples{
commonSuffix( c( "ABCDE", "CDE", "cDE" ))
#=> [1] "DE"

commonSuffix( c( "ABC", "abc", "def" ))
#=> [1] ""

commonSuffix( c( "ABCDE", "CDE", "" ))
#=> [1] ""

commonSuffix( c( "ABCDE", "ABCDE", NA ))
#=> [1] NA

commonSuffix( c( "A\u00dfC", "A\u00dfC", "a\u00dfC" ))
#=> [1] "ßC"

commonSuffix( c("ABCDE", "CDE", "cDE" ), ignoreCase= TRUE )
#=> [1] "cde"

}
\seealso{
\code{\link{commonPrefix}}
}
