% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtils.R
\name{expect_equalOrNull}
\alias{expect_equalOrNull}
\title{Extended equality tests}
\usage{
expect_equalOrNull(got, want, label = NULL, info = NULL, ...)
}
\arguments{
\item{got}{The object to test.}

\item{want}{The expected value, may be \code{NULL}.}

\item{label}{Object label.}

\item{info}{Extra info.}

\item{...}{Extra parameters passed to \code{\link{all.equal}}.}
}
\value{
Called only for its side effects during testing.
}
\description{
This general test fixture extends \code{expect_equal} to support \code{NULL}.
}
\examples{
\dontrun{
### Pass
expect_equalOrNull( 1, 1 )
expect_equalOrNull( list( a= 1,b= 2 ), list( b= 2, a= 1 ))
x <- list( a= 1, b= 2)["c"]
expect_equalOrNull( x, NULL )

### Fail
expect_equalOrNull( 1, "1" )
expect_equalOrNull( 1, NULL )
expect_equalOrNull( NULL, 1 )
}
}

