% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{fixNull}
\alias{fixNull}
\title{Replace a null value}
\usage{
fixNull(x, val = NULL, func = NULL, ...)
}
\arguments{
\item{x}{Check this to see if this is \code{NULL}.}

\item{val}{Return this value if \code{x} is \code{NULL}. Default is \code{NULL}.}

\item{func}{Called to generate and
return a replacement value, unless \code{val} is set, in which case that is
returned; any value returned by \code{func} is ignored. Can be a bare function
name, string, or function definition.}

\item{...}{Additional parameters to pass to \code{func}, if any. If any are
unnamed, \code{val} must be specified or explicitly set \code{NULL}. If \code{val} is not
specified, then due to R's parameter parsing, the first unnamed extra
parameter will be used as the value of \code{val} instead of being passed to the
function. That means it will also be returned as the NULL replacement
value. This "parameter swallowing" is a potential issues with all
functions using "..." with parameters that have default values. Handling of
this may change in future implementations.}
}
\value{
\code{x}, or if that is \code{NULL} then \code{val}, or if that is \code{NULL} the result
of calling \code{func} with any parameters \code{...}, or \code{NULL} if both \code{val} and
\code{func} are \code{NULL}. If both \code{val} and \code{func} are set, \code{val} will be
returned, but \code{func} will still be called for any potential side effects.
}
\description{
Returns the value \code{x}, unless x is \code{NULL}, in which case it returns either
\code{val} or, if \code{val} is NULL (the default), it returns the result of calling
the function \code{func}. Additional \code{...} parameters will be passed to \code{func}
if called. If both a replacement value and function are specified, the
replacement value is returned but the function is still called for its side
effects. Its return value will be ignored. This can be used e.g.
with a "warning" or a logging function to report when a replacement occurs.
Leaving both \code{val} and \code{func} at their default \code{NULL} values means this
function will have no net effect, i.e. it will replace \code{NULL} with \code{NULL}.
Note, due to R's argument parsing, if you want to call \code{func} and pass it
unnamed parameters via \code{...}, \code{val= NULL} must be set.
}
\examples{
fixNull( 42, -1 )
#> 42

fixNull( NULL, -1 )
#> -1

fixNull( NULL )
#> NULL

\dontrun{
# Return `val`, calling function only for side effect.
# Val is specified, so no issue with unnamed parameters
fixNull( NULL, 42, func="warning", "Was NULL" )
#> 42
#> Warning message:
#> In fixNull( NULL, func="warning", "Failed!" ) : Was NULL
}

# Calling function (with extra parameter) to generate replacement value.
# Extra parameter is named, so no issue with unspecified `val`.
fixNull( NULL, func= function( num ) { num * log(num) }, num= 42 )
#> [1] 156.9821

# Calling function (with extra parameters) to generate replacement value.
# Some extra  parameters are unnamed, so must specify `val= NULL`.
fixNull( NULL, val=NULL, func= paste, "this", "that", sep= "-" )
#> [1] "this-that"
}
