% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{githubGet}
\alias{githubGet}
\title{Download a file from GitHub}
\usage{
githubGet(file, repo, commit, dir = "./", force = FALSE, sha256 = NULL, ...)
}
\arguments{
\item{file}{The name of the file to download}

\item{repo}{The repository to download from, e.g. "Jefferys/JefferysRUtils"}

\item{commit}{The commit, tag, or branch to download.}

\item{dir}{The directory to download to, by default the current working directory.}

\item{force}{Download and over-write if already exists in output directory.}

\item{sha256}{Expected SHA256 checksum.}

\item{...}{Other parameters to pass through to download.file}
}
\value{
The exit code from download.file - 0 for success, positive integer
for error, and NA if not attempted due to an existing file.
}
\description{
Given the file name, repository, and commit/tag/branch, downloads the
specified (public) file from GitHub unless it already exists or is forced.
Does not support private repositories or directories.
}
\examples{
\dontrun{
githubGet( "README.Rmd", repo= Jefferys/JefferysRUtils,
           commit= 'e5dab82f022afa311be9d078d3e080fb88a2cfaf' )
githubGet( "README.Rmd", repo= Jefferys/JefferysRUtils,
           commit= 'dev' )
}
}
