% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%p\%}
\alias{\%p\%}
\alias{\%pp\%}
\title{Binary operator versions of paste and paste0}
\usage{
x \%p\% y

x \%pp\% y
}
\arguments{
\item{x}{The first object to paste}

\item{y}{The second object to paste}
}
\value{
Returns the result of pasting the two objects together:
\preformatted{
    x \%p\% y == paste0(x, y)
    x \%pp\% y == paste(x, y)
}
}
\description{
Like most operators, can split across lines with operator at front of second
line only if inside parenthesis. Unlike most operators, this is pretty
likely to happen. See Examples.
}
\examples{
"Hello, " \%p\% "world!" == "Hello, world!"
"Hello," \%pp\% "world!" == "Hello, world!"
name <- "Amy"
"Hello," \%pp\% name \%p\% "!" == "Hello, Amy!"
"Hello," \%pp\%
   "world!" == "Hello, world!"
("Hello,"
    \%pp\% "world!") == "Hello, world!"
# "Hello,"
#    \%pp\% "world!" == "Hello, world!"
# Error: unexpected SPECIAL in "     \%pp\%"

}

