% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{ifElseMore}
\alias{ifElseMore}
\title{An ifelse handling NA and NULL too.}
\usage{
ifElseMore(x, true = TRUE, false = FALSE, na = NA, null = NULL)
}
\arguments{
\item{x}{The value to test}

\item{true}{The value to return if \code{x} is \code{TRUE}. Returns \code{TRUE} by default.}

\item{false}{The value to return if \code{x} is \code{FALSE}. Returns \code{FALSE} by default.}

\item{na}{The value to return if \code{x} is \code{NA}. Returns \code{NA} by default.}

\item{null}{The value to return if \code{x} is \code{NULL}. Returns \code{NULL} by default.}
}
\value{
Depending on the value of x, returns the values of the matching
parameter \code{true}, \code{false}, \code{na}, or \code{null}.
}
\description{
Tests a Boolean scalar value and returns one of four values:
\itemize{
\item If \code{x} is \code{TRUE}, returns the value of the \code{true} parameter (TRUE by default).
\item If \code{x} is \code{FALSE}, returns the value of \code{false} parameter (FALSE by default).
\item If \code{x} is \code{NA}, returns the value of \code{na} parameter (NA by default).
\item If \code{x} is \code{NULL}, returns the value of \code{null} parameter (NULL by default).
}
}
\examples{
ifElseMore( 3 > 2 )
#> [1] TRUE
ifElseMore( 3 < 2 )
#> [1] FALSE
ifElseMore( 3 < NA )
#> [1] NA

ifElseMore( 3 > 2, "it's True", "it's False", "it's NA", "it's Null" )
#> [1] "it's True"
ifElseMore( 2 > 3, "it's True", "it's False", "it's NA", "it's Null" )
#> [1] "it's False
ifElseMore( NA > 2, "it's True", "it's False", "it's NA", "it's Null" )
#> [1] "it's NA"

x <- NULL
ifElseMore( x )
#> [1] NULL
ifElseMore( x, "it's True", "it's False", "it's NA", "it's Null" )
#> [1] "it's Null"

}
