% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initSayLoggers}
\alias{initSayLoggers}
\title{Initialize default loggers}
\usage{
initSayLoggers(file = packageName() \%p\% ".log", fileLevel = WARN,
  consoleLevel = INFO)
}
\arguments{
\item{file}{The name of the log file, by default \file{"\var{package}.log"}
where \var{package} is guessed using \code{link{packageName()}}.}

\item{fileLevel}{Only messages at least this important will be saved to the
log file, by default "WARN".}

\item{fileLevel}{Only messages at least this important will be printed to the
console, by default "INFO".}
}
\value{
Nothing, called only for its side effect of initializing loggers.
}
\description{
Sets up the default \code{"\var{package}.console"} and
\code{"\var{package}.file"} loggers. Use the \code{say\var{level}} function
to log simultaneously to them. If you log from a script or a console and
don't specify the log file, it will write to the probably hidden file
\file{".log"}. Allowed log levels are, in order of decreasing importance:
\code{OFF}, \code{FATAL}, \code{ERROR}, \code{WARN}, \code{INFO},
\code{DEBUG}, \code{TRACE}.
}
\examples{
\dontrun{
  initSayLoggers()
  # Won't be logged
  sayDebug('Initialized logging.')
  # Logged only to console, not file
  sayInfo('Welcome!')
  # Logged to console and file.
  sayWarn("This conversation is being monitored.")
}
}

