% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{isDuplicated}
\alias{isDuplicated}
\title{Is a value duplicated?}
\usage{
isDuplicated(x, ...)
}
\arguments{
\item{x}{The object to evaluate for duplicates. Handled generically so can
be vector-like (values), data.frame-like (rows), or matrix-like (rows, or
other dimension if specified via \verb{MARGIN=<INT>} ).}

\item{...}{Extra parameters like \code{MARGIN} for the appropriate \verb{duplicated.<x>()}
implementation, except for \code{fromLast}, which is used internally and will
result in an error.}
}
\value{
Logical vector indicating for each element of x if it is a repeated
value. For data frames, one element per row. For matrix and arrays, what
is returned depends on MARGIN and how degenerate your matrix or array is.
}
\description{
Checks each value in a vector, data frame, matrix, array, etc,  to see if it
repeats. If it does, it and all its repeated values are reported as
duplicates. \code{NA} and \code{NaN} are treated as values that can be duplicated or
not; to consider them (and potentially \code{""}, and/or \code{NaN}, etc) as not
duplicates, use \verb{incomparables=}.
}
\details{
Basically a simple wrapper around the algorithm:
\code{is.duplicated(x) == duplicated(x) | duplicated(x, fromLast = TRUE)}
and replicates the potentially inconsistent behavior of duplicated with
respect to matrices and arrays, especially degenerate matrices and arrays.
}
\examples{
isDuplicated( c( 1, 2, 3, 4, 3, 1 ))
#> [1]  TRUE FALSE  TRUE FALSE  TRUE  TRUE

mat <- matrix( c( 1, 3,  1,
                 NA, 3, NA,
                  2, 3,  2 ), nrow= 3, byrow= TRUE )
isDuplicated( mat, MARGIN= 2 )
#> [1]  TRUE FALSE  TRUE

}
\seealso{
\code{duplicated()}
}
