% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JefferysRUtils-package.R
\docType{package}
\name{JefferysRUtils-package}
\alias{JefferysRUtils}
\alias{JefferysRUtils-package}
\alias{JeffersRUtils-package}
\title{JefferysRUtils: My bag of tricks for R coding}
\description{
A collection of utilities that I use in R code. Each is too small to warrant their own package, they are relatively unrelated, and depending on them all as separate utility packages is too many dependencies for now. May split some out later, so be warned.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/jefferys/JefferysRUtils}
}

}
\keyword{internal}
