% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.list}
\alias{merge.list}
\title{Merge two lists}
\usage{
\method{merge}{list}(x, y, keepOrder = FALSE)
}
\arguments{
\item{x}{A list}

\item{y}{Another list}

\item{keepOrder}{How to order list when dropping duplicate elements from
\code{y} If FALSE, the default, just drops elements from \code{x} with the
same name as an element in \code{y}, before appending \code{y}. If TRUE,
preserves the order of element names as much as possible, by replacing the
value of duplicated element names with those from \code{y}, then appending
\code{y} with duplicated elements dropped.}
}
\value{
A list consisting of all elements from \code{x} not also in \code{y},
  and then all elements in \code{y}, or
}
\description{
Merges two lists, appending the second to the first. Expects that all
elements are named and that within a list, all names are unique. If the same
name is used in both lists, the element (value) from the second list is kept,
the one from the first list is dropped. That results in the order of elements
in the merged list being different. If \code{keepOrder = TRUE} is set, then
the order of elements in the first list is kept with the duplicated elements
from the second dropped.
}

