% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringUtils.R
\name{revString}
\alias{revString}
\title{Reverse a string}
\usage{
revString(x)
}
\arguments{
\item{x}{A vector of strings to reverse. Returns NA for NA and empty for empty.
If x=NULL or character(0), returns character(0).}
}
\value{
A vector of strings, each reversed. This is utf8 aware.
}
\description{
Reverse a string
}
\examples{
revString( "ABC" )
#=> revString( "ABC" )

revString( c( "ABC", "ABC", "A\u00dfB", 123, "", "x", NA ))
#=> [1] "CBA" "CBA" "BßA" "321" ""    "x"   NA

revString( NULL )
#=> character(0)

revString( character(0) )
#=> character(0)

}
