% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{saveFileSum}
\alias{saveFileSum}
\title{Create a checksum file}
\usage{
saveFileSum(
  file,
  algo = c("md5", "sha1", "crc32", "sha256", "sha512"),
  checkFile = NULL,
  show = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{file}{The file to validate}

\item{algo}{The checksum algorithm to use (see \code{digest::digest()}).}

\item{checkFile}{The name of the checksum file to create.
By default named the same as file, with an additional extension matching
\code{algo}, e.g. for an md5 checkFile for \code{file.txt}, the default would be
\code{file.txt.md5}. If this filename includes a directory component, it will
be created if needed.}

\item{show}{Should the checkFiles be visible (i.e. NOT start with a ".")? This is
ignored if the checkFile is specified. Note, if the file is
hidden, then its default checksum file will be hidden regardless of this setting.}

\item{overwrite}{If the checkFile already exists and does not match, it is
an error unless this is set TRUE, in which case the file is just overwritten.
If it matches, then done (in this case modify timestamps will not be changed).}
}
\value{
The generated checkFile
}
\description{
Calculates a checksum for a file and saves it to file that by default is
named the same, except with an appended extension matching the algorithm.
}
