% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{slurp}
\alias{slurp}
\title{Read a file's lines in as vector}
\usage{
slurp(file, commentString = NULL, skipLines = 0)
}
\arguments{
\item{file}{The file to read}

\item{commentString}{If specified, any line beginning with this comment
string will be dropped (leading whitespace is ignored). This is parsed as a
regualr expression, but that shouldn't be a problem for most normal line
comment strings (\code{"#"}, \code{"##"}, \code{"//"}, \code{";"}), leading
whitespace is ignored. If null or empty, will be ignored. Backslashes
probably need to be doubled.}

\item{skipLines}{If specified, skip this many lines at the start of the file.
This will be done after comments are filtered, so don't skip lines if
they are also comments.}
}
\value{
The file's content as a string vector, one element per line, including
  empty lines as empty strings, but excluding filtered lines. If no lines are
  left after dropping comments and skipping the headers, will return
  \code{character(0)}
}
\description{
Essentially a convenience wrapper around \code{\link{readLines}}. Reads a
file's lines into a string vector. Can skip comments if specify a comment
string, and drops the header line if told to. Inspired by perl's "slurp"
mode.
}
\section{Warnings}{


\describe{
   \item{
      \command{Skipping more lines, \var{skip}, than lines to skip, \var{lines}.}
   }{
      There are fewer lines left (after skipping comments) than you asked to
      skip. Will return \code{character(0)} to indicate no kept lines/strings.
   }
}
}

