% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{useResultsDir}
\alias{useResultsDir}
\title{Make a directory for results.}
\usage{
useResultsDir(base, dir, rerun = FALSE)
}
\arguments{
\item{base}{The base directory to create a results directory in. Will be
created (recursively) if needed, unless \code{rerun= TRUE} is set. May be
absolute or relative.}

\item{dir}{The results directory under \verb{base=} to be created (recursively).
Must not exist unless \code{rerun= TRUE} is set. Must be relative to \verb{base=}.}

\item{rerun}{Set \code{TRUE} to reuse an existing run directory; it must already
exist.}
}
\value{
The path to the created (or existing) output directory,
\verb{<base>/<outDir>}.
}
\description{
Make a directory for results.
}
\examples{
baseDir <- tempfile()
outDirName <- "run-1"
dir.exists( file.path( baseDir, outDirName )) == FALSE
myOutDir <- useResultsDir( baseDir, outDirName )
dir.exists( file.path( baseDir, outDirName )) == TRUE
useResultsDir( baseDir, outDirName, rerun=TRUE ) == myOutDir

}
