% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JefferysRUtils.R
\docType{package}
\name{JefferysRUtils}
\alias{JefferysRUtils}
\title{JefferysRUtils: My bag of tricks for R coding}
\description{
A collection of utilities that I use in R code. Each is too small to
warrant their own package, they are relatively unrelated, and depending on
them all as separate utility packages is too many dependencies for now. May
split some out later, so be warned.
These are useful for me. Your mileage may vary. Some are unfinished
and possibly broken
}
